remove the level cap and show equipment bonuses in character details
Two related pieces of progression work.

No level cap. Canary has none — uint32_t level, uint64_t experience, and no
CHECK on any of it in schema.sql — and every limit here was ours. Only one
was real: the experience math ran in JS number, exact to 2^53, so past level
~81456 levels silently resolved wrong. Experience is now bigint end to end
(Canary's own P/6*100 order, verified identical to the previous P*100/6 for
every level to 200000) and serialised as a decimal string on the wire, since
JSON has no bigint. Migration 072 drops the level, experience, health, and
mana upper bounds, leaving column widths as the only limit.
MAX_STORABLE_CHARACTER_LEVEL = 821009 replaces MAX_CHARACTER_LEVEL: not a
rule, just where the experience stops fitting a signed 64-bit column.

Migration 071 first: 007 capped stored health and mana at 100000, which the
50000-level raise left behind. A level-5000 sorcerer (150025 max mana) then
failed every persist on characters_mana_upper_bound, surfacing in play as
"potion persist failed" and a cache resync.

Equipment bonuses reach the panel. Item skill and magic-level modifiers were
applied by combat but the panel projected the unequipped values, and the item
`speed` attribute (boots of haste, time ring) was never applied to movement at
all — only imbuement Swiftness fed the stat modifier. Skills now go through
one shared equipmentSkillModifier that playerCombatSkill also calls, so the
number shown and the number swung cannot drift. Bonused stats render tinted
and break down on hover as base + equipment via a HoverTooltip extracted from
ProgressionBar, which fits itself inside the scrolling panel on both axes.

+++ server/src/playtest/scenarios/weaponParity.ts
@@ -176,7 +176,8 @@ try {
   );
   check(
     "melee-experience-progression",
-    knight.progression.experience - xpBefore >= melee.kills * 40,
+    BigInt(knight.progression.experience) - BigInt(xpBefore) >=
+      BigInt(melee.kills * 40),
     `experience ${xpBefore} -> ${knight.progression.experience}`,
   );
   knight.client.terminate();
